requirements.txt
livekit-plugins-google
livekit-plugins-noise-cancellation
duckduckgo-search
requests
langchain_community
livekit-agents==1.2.3
livekit-plugins-openai>=0.6.0
livekit-plugins-elevenlabs>=0.5.0
livekit-plugins-deepgram>=0.6.0
livekit-plugins-silero>=0.5.0
python-dotenv==1.0.0

# Additional Dependencies
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0